module-5: implemented command handlers with cqs for use cases

+++ src/main/java/app/delivery/core/domain/courier/aggregate/Transport.java
@@ -21,7 +21,7 @@ public enum Transport {
     private final int capacity;
 
 
-    public boolean canCarry(int weight) {
+    public boolean canCarry(double weight) {
         return capacity >= weight;
     }
